Change syntax to arrow
