+++ src/config.ts
@@ -1,0 +1,9 @@
+export function config() {
+  return {
+    version: '0.1',
+    backend: {
+      user: 'jdoe',
+      password: 'password'
+    }
+  };
+}
